pyproject.toml
[project]
name = "webapp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "django>=5.2.5",
    "django-nested-admin>=4.1.4",
    "djangorestframework>=3.16.1",
]
